fix: Change email subject line to use short month format
- Update subject line format from 'November 2, 2025' to 'Nov 2, 2025'
- Subject line now: 'Weekly Time Tracking Report - Nov 2, 2025 - Nov 8, 2025'

+++ app/lib/services/email.service.ts
@@ -243,12 +243,12 @@ export async function sendWeeklyReportEmail(
     const fromEmail = options.from || process.env.RESEND_FROM_EMAIL || 'noreply@flexscale.com';
     const replyTo = options.replyTo || fromEmail;
 
-    // Format dates for subject line to match the period format in email body
+    // Format dates for subject line (e.g., "Nov 2, 2025")
     const formatDateForSubject = (dateString: string) => {
       const date = new Date(dateString);
       return date.toLocaleDateString('en-US', { 
         year: 'numeric', 
-        month: 'long', 
+        month: 'short', 
         day: 'numeric' 
       });
     };
